Adding simple test code that can read rule input/output files.
A single Generator class handles all three "types" of Generators.
The parameters used by the Generator class are relatively simple, the
extra complexity of an entire inheritance hierarchy doesn't seem to be
worth it...

We use a 2-pass algorithm when reading the input files to make it easier
to handle the different orderings that the points and weights can be in.
Sometimes all the weights are first, sometimes the points and weights are
interleaved, sometimes the points come before the weights... We can now
handle all of these cases.