Experimenting with different sinusoidal alpha values.
Adding an alpha which maximizes I(f), although this seems to be a very
easy function to integrate, and the 3rd and 5th-order rules have very
low effectivity for it.

Picking large alpha which is not a multiple of pi is also not
interesting, since the H1 norms are huge, the normalized error also
appears to be very small.

Write sinusoidal output files with the value of alpha.