Set xtol, ftol, maxeval in only one place.
I also experimented with different values of these, but if you make
them too loose, there are tons of false positives and very few
"actually converged" results.

Also set maxeval on the "local" optimizer, as I had not previously
done that before. This may be a good way to experiment with how
much time is spent on any one optimization attempt.