Add Python script for degree=3 rule with 2 edge orbits.
I believe we can show analytically that there is no solution to this
system of equations, and the Python script agrees with this, but that
could just be due to a bug in the Python script... I do find it
interesting that the rule with all median orbits has infinitely many
solutions while the rule with all edge orbits has no solutions.
In the end, none of these rules are optimal rules that would end
up being tabulated, but they are interesting test cases nonetheless.