One of the 'solutions' that we wrote down for degree 3 with 6 QP was wrong.
This one also behaved strangely in nlopt; when used as an initial
guess it did not converge to itself... since this is a weird system of
equations I can't rule out this kind of behavior... it's possible
there are spurious solutions? The other two rules in this
configuration that we had recorded seemed fine. Luckily this
configuration has an "infinite" number of solutions so we could just
replace the incorrect one with a different one that does work.

* Fix incorrect ordering in degree=3 1,1,1,0,0 rule. This was simply a
  transcription error.
* Just keep three of the many converged degree=3 solutions.