Fix plot of weights for the NI rules.
At first I was confused because I kept thinking that min_alpha was in
the middle of the range of alpha values, but of course it is not. In
the end, I figured out I could use the min_root_x1 values for the
second half of the plot, just like I had done for the (alpha,
x1(alpha)) plot.